crates/opencassava-core/src/transcription/parakeet_requirements.txt
nemo_toolkit[asr]>=2.0.0
soundfile>=0.12.0
numpy>=1.26.0
